Actually login with the user info

+++ login-server/src/main/java/work/fking/pangya/login/model/PlayerAccount.java
@@ -3,7 +3,7 @@ package work.fking.pangya.login.model;
 import java.time.LocalDateTime;
 import java.util.Arrays;
 
-public record PlayerAccount(long id, String username, String nickname, byte[] passwordHash, Status status, LocalDateTime suspensionLiftTimestamp) {
+public record PlayerAccount(int id, String username, String nickname, byte[] passwordHash, Status status, LocalDateTime suspensionLiftTimestamp) {
 
     public enum Status {
         ACTIVE,
